+++ database/migrations/2024_10_29_013348_create_dashboards_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,16 +12,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('dashboards', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('name');
-            $table->integer('column_count')->default(2);
-            $table->string('app')->nullable();
-            $table->boolean('is_default')->default(false);
-            $table->boolean('is_starred')->default(false);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('dashboards')) {
+            Schema::create('dashboards', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('name');
+                $table->integer('column_count')->default(2);
+                $table->string('app')->nullable();
+                $table->boolean('is_default')->default(false);
+                $table->boolean('is_starred')->default(false);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +31,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('dashboards');
+        if (Schema::hasTable('dashboards')) {
+            $count = DB::table('dashboards')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('dashboards');
+            }
+        }
     }
 };
