+++ database/migrations/2024_10_29_013417_create_widgets_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,15 +12,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('widgets', function (Blueprint $table) {
-            $table->id();
-            $table->string('type');
-            $table->integer('column');
-            $table->integer('order')->default(0);
-            $table->foreignId('dashboard_id')->constrained()->onDelete('cascade');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('widgets')) {
+            Schema::create('widgets', function (Blueprint $table) {
+                $table->id();
+                $table->string('type');
+                $table->integer('column');
+                $table->integer('order')->default(0);
+                $table->foreignId('dashboard_id')->constrained()->onDelete('cascade');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,6 +30,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('widgets');
+        if (Schema::hasTable('widgets')) {
+            $count = DB::table('widgets')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('widgets');
+            }
+        }
     }
 };
