+++ database/migrations/2024_10_29_013519_create_projects_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,13 +12,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('projects', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('identifier');
-            $table->json('enabledModules');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('projects')) {
+            Schema::create('projects', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('identifier');
+                $table->json('enabledModules');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -25,6 +28,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('projects');
+        if (Schema::hasTable('projects')) {
+            $count = DB::table('projects')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('projects');
+            }
+        }
     }
 };
